Latest unix commands to extract data from corpus
These two text files are able to extract the titles, or the titles and subtitles AND create new .txt files in the 'corpus_output' directory. This is an improvement on the last iteration because it creates the EQUIVALENT amount of .txt files so Voyant is able to analyse information such as average words per article, etc.